+++ examples/src/donate_to_treasury_example.c
@@ -1,0 +1,162 @@
+/**
+ * \file donate_to_treasury_example.c
+ *
+ * \author angel.castillo
+ * \date   Aug 18, 2025
+ *
+ * Copyright 2025 Biglup Labs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License")
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/* INCLUDES ******************************************************************/
+
+#include "providers/provider_factory.h"
+
+#include "utils/console.h"
+#include "utils/utils.h"
+#include <cardano/cardano.h>
+
+#include <stdlib.h>
+#include <string.h>
+
+/* CONSTANTS *****************************************************************/
+
+static const char*    SERIALIZED_BIP32_KEY_HANDLER = "0a0a0a0a01010000005c97db5e09b3a4919ec75ed1126056241a1e5278731c2e0b01bea0a5f42c22db4131e0a4bbe75633677eb0e60e2ecd3520178f85c7e0d4be77a449087fe9674ee52f946b07c1b56d228c496ec0d36dd44212ba8af0f6eed1a82194dd69f479c603";
+static const uint64_t CONFIRM_TX_TIMEOUT_MS        = 240000U;
+static const int64_t  LOVELACE_TO_DONATE           = 2000000U;
+static const uint64_t PAYMENT_CRED_INDEX           = 0U;
+static const uint64_t STAKE_CRED_INDEX             = 0U;
+static const uint64_t SECONDS_IN_TWO_HOURS         = 60UL * 60UL * 2UL;
+
+static const cardano_account_derivation_path_t ACCOUNT_DERIVATION_PATH = {
+  .purpose   = 1852U | 0x80000000,
+  .coin_type = 1815U | 0x80000000,
+  .account   = 0U
+};
+
+static const cardano_derivation_path_t SIGNER_DERIVATION_PATH = {
+  .purpose   = 1852U | 0x80000000,
+  .coin_type = 1815U | 0x80000000,
+  .account   = 0U,
+  .role      = 0U,
+  .index     = 0U
+};
+
+/* DECLARATIONS **************************************************************/
+
+/**
+ * \brief Retrieves the password for the secure key handler.
+ *
+ * \param buffer The buffer where to write the password.
+ * \param buffer_len The size of the buffer.
+ *
+ * \return The length of the password (or -1 on error).
+ */
+static int32_t
+get_passphrase(byte_t* buffer, const size_t buffer_len)
+{
+  console_warn("Enter passphrase: ");
+  char          password[128] = { 0 };
+  const int32_t password_len  = console_read_password(password, sizeof(password));
+
+  if (buffer_len < password_len)
+  {
+    return -1;
+  }
+
+  cardano_utils_safe_memcpy(buffer, buffer_len, &password[0], password_len);
+
+  // Clear local password from memory
+  cardano_memzero(password, sizeof(password));
+
+  return password_len;
+}
+
+/* MAIN **********************************************************************/
+
+/**
+ * \brief Entry point of the program.
+ *
+ * \return Returns `0` on successful execution, or a non-zero value if there is an error.
+ */
+int
+main(void)
+{
+  console_info("Donate lovelace to treasury Example");
+  console_info("libcardano-c:  V-%s\n", cardano_get_lib_version());
+
+  console_info("This example will donate %d lovelace to the treasury.", LOVELACE_TO_DONATE);
+
+  console_set_foreground_color(CONSOLE_COLOR_GREEN);
+  console_write("\nUse passphrase: 'password'\n\n");
+  console_reset_color();
+
+  const char* api_key = getenv("BLOCKFROST_API_KEY");
+
+  if (api_key == NULL)
+  {
+    console_error("BLOCKFROST_API_KEY environment variable is not set.\n");
+
+    return EXIT_FAILURE;
+  }
+
+  // 0.- Initialize dependencies
+  cardano_secure_key_handler_t*  key_handler     = create_secure_key_handler(SERIALIZED_BIP32_KEY_HANDLER, cardano_utils_safe_strlen(SERIALIZED_BIP32_KEY_HANDLER, 256), get_passphrase);
+  cardano_provider_t*            provider        = create_provider(CARDANO_NETWORK_MAGIC_PREPROD, api_key);
+  cardano_address_t*             payment_address = create_address_from_derivation_paths(key_handler, ACCOUNT_DERIVATION_PATH, PAYMENT_CRED_INDEX, STAKE_CRED_INDEX);
+  cardano_utxo_list_t*           utxo_list       = get_unspent_utxos(provider, payment_address);
+  cardano_protocol_parameters_t* protocol_params = get_protocol_parameters(provider);
+
+  // 2 hours from now in UNIX time (seconds)
+  const uint64_t invalid_after = cardano_utils_get_time() + SECONDS_IN_TWO_HOURS;
+
+  // 1.- Build transaction
+  console_info("Building transaction...");
+
+  cardano_tx_builder_t* tx_builder = cardano_tx_builder_new(protocol_params, provider);
+
+  cardano_tx_builder_set_utxos(tx_builder, utxo_list);
+  cardano_tx_builder_set_change_address(tx_builder, payment_address);
+  cardano_tx_builder_set_invalid_after_ex(tx_builder, invalid_after);
+  cardano_tx_builder_set_donation(tx_builder, LOVELACE_TO_DONATE);
+
+  cardano_transaction_t* transaction = NULL;
+  cardano_error_t        result      = cardano_tx_builder_build(tx_builder, &transaction);
+
+  if (result != CARDANO_SUCCESS)
+  {
+    console_error("Failed to build transaction");
+    console_error("Error [%d]: %s", result, cardano_error_to_string(result));
+    console_error("%s", cardano_tx_builder_get_last_error(tx_builder));
+
+    return result;
+  }
+
+  // 2.- Sign transaction
+  sign_transaction(key_handler, SIGNER_DERIVATION_PATH, transaction);
+
+  // 3.- Submit transaction & confirm
+  submit_transaction(provider, CONFIRM_TX_TIMEOUT_MS, transaction);
+
+  // Cleanup
+  cardano_provider_unref(&provider);
+  cardano_address_unref(&payment_address);
+  cardano_utxo_list_unref(&utxo_list);
+  cardano_protocol_parameters_unref(&protocol_params);
+  cardano_tx_builder_unref(&tx_builder);
+  cardano_transaction_unref(&transaction);
+  cardano_secure_key_handler_unref(&key_handler);
+
+  return EXIT_SUCCESS;
+}